Create groups in App level rather than in ContentGrid.
